modules/PythonScaler/requirements.txt
juju
aiohttp
mysql-connector
